BANGER code to run praise over players based on opencv
This is a big change to how we detect what players to praise. Before, we entered a static number of iterations. We now orientate to the individual training location, as before, we then use a reference image that is a left-side snippet of the "good" training ratings, we direct the script to click, and reuse our praise button logic, and iterate.

Script did not work at first, since it was continuously clicking on the same button multiple times. We used "grouping" to infer where we found singular images, and then iterate through them all. Limitation is still around it only being able to do what is in a screenshot. May need to see if we can suss out how to do a full page, the "page down" and keep going. As this is color based, if a player was already praised, would may be able to do this later.